change year column order

+++ src/aclimate_resampling/resampling.py
@@ -536,7 +536,7 @@ class Resampling():
       columns.remove('year')
       new_columns = columns[:2] + ['year'] + columns[2:]
       df = df[new_columns]
-      df['year'] = df['year'].astype('int')
+      
 
       #df.groupby(['year', 'month', 'day']).max().reset_index().to_csv(os.path.join(output_summary, f"{station}_ejemplo.csv"))
